Moved running notes list to running notes components and added note about no notes matching filter
